Moved user package to the model package
